@@ -1,5 +1,5 @@
-﻿// Copyright (c) 2019-2021 ReactiveUI Association Inc. All rights reserved.
-// ReactiveUI Association Inc licenses this file to you under the MIT license.
+﻿// Copyright (c) 2019-2021 ReactiveUI Association Incorporated. All rights reserved.
+// ReactiveUI Association Incorporated licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for full license information.
 
 using System;
@@ -13,7 +13,7 @@ namespace ReactiveMarbles.NuGet.Helpers
     /// </summary>
     public class FilesGroup
     {
-        private readonly DirectoryNode _rootNode = new DirectoryNode(string.Empty);
+        private readonly DirectoryNode _rootNode = new(string.Empty);
 
         /// <summary>
         /// Gets for a file name, the nearest matching full name in the shallowest of the hierarchy.
@@ -95,10 +95,10 @@ namespace ReactiveMarbles.NuGet.Helpers
 
         private class DirectoryNode : IEqualityComparer<DirectoryNode>, IComparable<DirectoryNode>
         {
-            private readonly Dictionary<string, DirectoryNode> _childNodesDict = new Dictionary<string, DirectoryNode>();
-            private readonly List<DirectoryNode> _childNodes = new List<DirectoryNode>();
-            private readonly List<FileNode> _files = new List<FileNode>();
-            private readonly Dictionary<string, FileNode> _filesDict = new Dictionary<string, FileNode>();
+            private readonly Dictionary<string, DirectoryNode> _childNodesDict = new();
+            private readonly List<DirectoryNode> _childNodes = new();
+            private readonly List<FileNode> _files = new();
+            private readonly Dictionary<string, FileNode> _filesDict = new();
 
             public DirectoryNode(DirectoryNode? parent, string name)
             {
