@@ -1,5 +1,5 @@
-﻿// Copyright (c) 2019-2021 ReactiveUI Association Inc. All rights reserved.
-// ReactiveUI Association Inc licenses this file to you under the MIT license.
+﻿// Copyright (c) 2019-2021 ReactiveUI Association Incorporated. All rights reserved.
+// ReactiveUI Association Incorporated licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for full license information.
 
 using System.Collections.Generic;
@@ -9,7 +9,7 @@ using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 
-using static Microsoft.CodeAnalysis.CSharp.SyntaxFactory;
+using static ReactiveMarbles.ObservableEvents.SourceGenerator.SyntaxFactoryHelpers;
 
 namespace ReactiveMarbles.ObservableEvents.SourceGenerator.EventGenerators.Generators
 {
@@ -20,15 +20,15 @@ namespace ReactiveMarbles.ObservableEvents.SourceGenerator.EventGenerators.Gener
     /// </summary>
     internal static class DelegateGenerator
     {
-        private static readonly QualifiedNameSyntax _subjectNamespace = QualifiedName(IdentifierName("Pharmacist"), IdentifierName("Common"));
-        private static readonly GenericNameSyntax _subjectType = GenericName(Identifier("SingleAwaitSubject"));
+        private static readonly QualifiedNameSyntax _subjectNamespace = QualifiedName("Pharmacist", "Common");
+        private static readonly GenericNameSyntax _subjectType = GenericName("SingleAwaitSubject");
 
         /// <summary>
         /// Generate our namespace declarations. These will contain our helper classes.
         /// </summary>
         /// <param name="declarations">The declarations to add.</param>
         /// <returns>An array of namespace declarations.</returns>
-        internal static IEnumerable<NamespaceDeclarationSyntax> Generate(IEnumerable<(INamedTypeSymbol TypeDefinition, bool IsAbstract, IEnumerable<IMethodSymbol> Methods)> declarations)
+        internal static IEnumerable<NamespaceDeclarationSyntax> Generate(IEnumerable<(INamedTypeSymbol TypeDefinition, bool IsAbstract, IReadOnlyCollection<IMethodSymbol> Methods)> declarations)
         {
             foreach (var groupedDeclarations in declarations.GroupBy(x => x.TypeDefinition.ContainingNamespace.Name).OrderBy(x => x.Key))
             {
@@ -39,8 +39,7 @@ namespace ReactiveMarbles.ObservableEvents.SourceGenerator.EventGenerators.Gener
 
                 if (members.Count > 0)
                 {
-                    yield return NamespaceDeclaration(IdentifierName(namespaceName))
-                        .WithMembers(List<MemberDeclarationSyntax>(members));
+                    yield return NamespaceDeclaration(namespaceName, members, true);
                 }
             }
         }
@@ -52,24 +51,25 @@ namespace ReactiveMarbles.ObservableEvents.SourceGenerator.EventGenerators.Gener
         /// <param name="isAbstract">If the delegates are abstract.</param>
         /// <param name="methods">The methods to generate delegate overloads for.</param>
         /// <returns>The generated class declarations.</returns>
-        private static ClassDeclarationSyntax GenerateClass(INamedTypeSymbol typeDefinition, bool isAbstract, IEnumerable<IMethodSymbol> methods)
+        private static ClassDeclarationSyntax GenerateClass(INamedTypeSymbol typeDefinition, bool isAbstract, IReadOnlyCollection<IMethodSymbol> methods)
         {
-                var modifiers = typeDefinition.IsAbstract || isAbstract
-                    ? TokenList(Token(SyntaxKind.PublicKeyword), Token(SyntaxKind.AbstractKeyword), Token(SyntaxKind.PartialKeyword))
-                    : TokenList(Token(SyntaxKind.PublicKeyword), Token(SyntaxKind.PartialKeyword));
-                return ClassDeclaration(typeDefinition.Name + "Rx")
-                    .WithModifiers(modifiers)
-                    .WithMembers(List(GenerateObservableMembers(methods)))
-                    .WithBaseList(BaseList(SingletonSeparatedList<BaseTypeSyntax>(SimpleBaseType(IdentifierName(typeDefinition.GenerateFullGenericName())))))
-                    .WithLeadingTrivia(XmlSyntaxFactory.GenerateSummarySeeAlsoComment("Wraps delegates events from {0} into Observables.", typeDefinition.GetArityDisplayName()))
-                    .WithObsoleteAttribute(typeDefinition);
+            var modifiers = typeDefinition.IsAbstract || isAbstract
+                ? new[] { SyntaxKind.PublicKeyword, SyntaxKind.AbstractKeyword, SyntaxKind.PartialKeyword }
+                : new[] { SyntaxKind.PublicKeyword, SyntaxKind.PartialKeyword };
+
+            var baseClasses = new[] { SimpleBaseType(typeDefinition.GenerateFullGenericName()) };
+
+            var attributes = RoslynHelpers.GenerateObsoleteAttributeList(typeDefinition);
+
+            return ClassDeclaration(typeDefinition.Name + "Rx", attributes, modifiers, GenerateObservableMembers(methods).ToList(), baseClasses, 1)
+                .WithLeadingTrivia(XmlSyntaxFactory.GenerateSummarySeeAlsoComment("Wraps delegates events from {0} into Observables.", typeDefinition.GetArityDisplayName()));
         }
 
-        private static IEnumerable<MemberDeclarationSyntax> GenerateObservableMembers(IEnumerable<IMethodSymbol> methods)
+        private static IEnumerable<MemberDeclarationSyntax> GenerateObservableMembers(IReadOnlyCollection<IMethodSymbol> methods)
         {
-            var methodDeclarations = new List<MethodDeclarationSyntax>();
-            var fieldDeclarations = new List<FieldDeclarationSyntax>();
-            var propertyDeclarations = new List<PropertyDeclarationSyntax>();
+            var methodDeclarations = new List<MethodDeclarationSyntax>(methods.Count);
+            var fieldDeclarations = new List<FieldDeclarationSyntax>(methods.Count);
+            var propertyDeclarations = new List<PropertyDeclarationSyntax>(methods.Count);
 
             foreach (var method in methods.OrderBy(y => y.Name))
             {
@@ -90,13 +90,12 @@ namespace ReactiveMarbles.ObservableEvents.SourceGenerator.EventGenerators.Gener
         /// <returns>The property declaration.</returns>
         private static PropertyDeclarationSyntax GeneratePropertyDeclaration(string observableName, IMethodSymbol method)
         {
+            var modifiers = new[] { SyntaxKind.PublicKeyword };
+            var attributes = RoslynHelpers.GenerateObsoleteAttributeList(method);
+
             // Produces:
             // public System.IObservable<type> MethodNameObs => _observableName;
-            return PropertyDeclaration(method.GenerateObservableTypeArguments().GenerateObservableType(), Identifier(method.Name + "Obs"))
-                .WithModifiers(TokenList(Token(SyntaxKind.PublicKeyword)))
-                .WithObsoleteAttribute(method)
-                .WithExpressionBody(ArrowExpressionClause(IdentifierName(observableName)))
-                .WithSemicolonToken(Token(SyntaxKind.SemicolonToken))
+            return PropertyDeclaration(method.GenerateObservableTypeArguments().GenerateObservableType(), method.Name + "Obs", attributes, modifiers, ArrowExpressionClause(IdentifierName(observableName)), 2)
                 .WithLeadingTrivia(XmlSyntaxFactory.GenerateSummarySeeAlsoComment("Gets an observable which signals when the {0} method is invoked.", method.ConvertToDocument()));
         }
 
@@ -111,25 +110,19 @@ namespace ReactiveMarbles.ObservableEvents.SourceGenerator.EventGenerators.Gener
             // Produces:
             // private readonly ReactiveUI.Events.SingleAwaitSubject<type> _methodName = new ReactiveUI.Events.SingleAwaitSubject<type>();
             var typeName = QualifiedName(_subjectNamespace, _subjectType.WithTypeArgumentList(method.GenerateObservableTypeArguments()));
+            var modifiers = new[] { SyntaxKind.PrivateKeyword, SyntaxKind.ReadOnlyKeyword };
+            var attributes = RoslynHelpers.GenerateObsoleteAttributeList(method);
 
-            return FieldDeclaration(VariableDeclaration(typeName)
-                .WithVariables(
-                    SingletonSeparatedList(
-                        VariableDeclarator(Identifier(observableName))
-                            .WithInitializer(
-                                EqualsValueClause(
-                                    ObjectCreationExpression(typeName).WithArgumentList(ArgumentList()))))))
-                .WithModifiers(TokenList(Token(SyntaxKind.PrivateKeyword), Token(SyntaxKind.ReadOnlyKeyword)));
+            var equalsClause = EqualsValueClause(ObjectCreationExpression(typeName));
+
+            return FieldDeclaration(typeName, observableName, equalsClause, attributes, modifiers, 2);
         }
 
         private static MethodDeclarationSyntax GenerateMethodDeclaration(string observableName, IMethodSymbol method)
         {
-            // Produces:
-            // /// <inheritdoc />
-            // public override void MethodName(params..) => _methodName.OnNext(...);
-            var methodBody = InvocationExpression(MemberAccessExpression(SyntaxKind.SimpleMemberAccessExpression, IdentifierName(observableName), IdentifierName("OnNext")));
-
-            var methodParameterList = method.GenerateMethodParameters();
+            var attributes = RoslynHelpers.GenerateObsoleteAttributeList(method);
+            var modifiers = new[] { SyntaxKind.PublicKeyword, SyntaxKind.OverrideKeyword };
+            IReadOnlyCollection<ArgumentSyntax> arguments;
 
             // If we have any members call our observables with the parameters.
             if (method.Parameters.Length > 0)
@@ -138,25 +131,27 @@ namespace ReactiveMarbles.ObservableEvents.SourceGenerator.EventGenerators.Gener
                 // If we have more than one parameter we have to pass them by value tuples, since observables only have one generic type parameter.
                 if (method.Parameters.Length == 1)
                 {
-                    methodBody = methodBody.WithArgumentList(method.Parameters[0].GenerateArgumentList());
+                    arguments = method.Parameters[0].GenerateArgumentList();
                 }
                 else
                 {
-                    methodBody = methodBody.WithArgumentList(method.Parameters.GenerateTupleArgumentList());
+                    arguments = method.Parameters.GenerateTupleArgumentList();
                 }
             }
             else
             {
-                methodBody = methodBody.WithArgumentList(RoslynHelpers.ReactiveUnitArgumentList);
+                arguments = RoslynHelpers.ReactiveUnitArgumentList;
             }
 
-            return MethodDeclaration(PredefinedType(Token(SyntaxKind.VoidKeyword)), method.Name)
-                .WithExpressionBody(ArrowExpressionClause(methodBody))
-                .WithParameterList(methodParameterList)
-                .WithObsoleteAttribute(method)
-                .WithModifiers(TokenList(Token(SyntaxKind.PublicKeyword), Token(SyntaxKind.OverrideKeyword)))
-                .WithLeadingTrivia(XmlSyntaxFactory.InheritdocSyntax)
-                .WithSemicolonToken(Token(SyntaxKind.SemicolonToken));
+            var methodParameterList = method.GenerateMethodParameters();
+
+            // Produces:
+            // /// <inheritdoc />
+            // public override void MethodName(params..) => _methodName.OnNext(...);
+            var methodBody = ArrowExpressionClause(InvocationExpression(MemberAccessExpression(SyntaxKind.SimpleMemberAccessExpression, observableName, "OnNext"), arguments));
+
+            return MethodDeclaration(attributes, modifiers, "void", method.Name, methodParameterList, 2, methodBody)
+                .WithLeadingTrivia(XmlSyntaxFactory.InheritdocSyntax);
         }
     }
 }
