@@ -1,5 +1,5 @@
-﻿// Copyright (c) 2019-2021 ReactiveUI Association Inc. All rights reserved.
-// ReactiveUI Association Inc licenses this file to you under the MIT license.
+﻿// Copyright (c) 2019-2021 ReactiveUI Association Incorporated. All rights reserved.
+// ReactiveUI Association Incorporated licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for full license information.
 
 using System.Collections.Generic;
@@ -10,7 +10,7 @@ using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 
-using static Microsoft.CodeAnalysis.CSharp.SyntaxFactory;
+using static ReactiveMarbles.ObservableEvents.SourceGenerator.SyntaxFactoryHelpers;
 
 namespace ReactiveMarbles.ObservableEvents.SourceGenerator.EventGenerators.Generators
 {
@@ -20,7 +20,7 @@ namespace ReactiveMarbles.ObservableEvents.SourceGenerator.EventGenerators.Gener
     internal abstract class EventGeneratorBase : IEventSymbolGenerator
     {
         /// <inheritdoc />
-        public abstract NamespaceDeclarationSyntax? Generate(INamedTypeSymbol item, bool generateEmpty);
+        public abstract NamespaceDeclarationSyntax? Generate(INamedTypeSymbol item);
 
         /// <summary>
         /// Generates an observable declaration that wraps a event.
@@ -42,22 +42,21 @@ namespace ReactiveMarbles.ObservableEvents.SourceGenerator.EventGenerators.Gener
             }
 
             var modifiers = eventDetails.IsStatic
-                ? TokenList(Token(SyntaxKind.PublicKeyword), Token(SyntaxKind.StaticKeyword))
-                : TokenList(Token(SyntaxKind.PublicKeyword));
+                ? new[] { SyntaxKind.PublicKeyword, SyntaxKind.StaticKeyword }
+                : new[] { SyntaxKind.PublicKeyword };
+
+            var attributes = RoslynHelpers.GenerateObsoleteAttributeList(eventDetails);
 
             // Produces for static: public static global::System.IObservable<(argType1, argType2)> EventName => (contents of expression body)
             // Produces for instance: public global::System.IObservable<(argType1, argType2)> EventName => (contents of expression body)
-            return PropertyDeclaration(observableEventArgType, prefix + eventDetails.Name)
-                .WithModifiers(modifiers)
-                .WithExpressionBody(expressionBody)
-                .WithSemicolonToken(Token(SyntaxKind.SemicolonToken))
-                .WithObsoleteAttribute(eventDetails)
-                .WithLeadingTrivia(XmlSyntaxFactory.GenerateSummarySeeAlsoComment("Gets an observable which signals when the {0} event triggers.", eventDetails.ConvertToDocument()));
+            return PropertyDeclaration(observableEventArgType, prefix + eventDetails.Name, attributes, modifiers, expressionBody, 2)
+                .WithLeadingTrivia(XmlSyntaxFactory.GenerateSummarySeeAlsoComment("Gets an observable which signals when the {0} event triggers.", eventDetails.ConvertToDocument()))
+                .WithSemicolonToken(Token(SyntaxKind.SemicolonToken));
         }
 
         private static (ArrowExpressionClauseSyntax ArrowClause, TypeSyntax EventArgsType) GenerateFromEventExpression(IEventSymbol eventSymbol, string dataObjectName)
         {
-            ArgumentListSyntax methodParametersArgumentList;
+            IReadOnlyCollection<ArgumentSyntax> methodParametersArgumentList;
             TypeSyntax eventArgsType;
 
             var invokeMethod = ((INamedTypeSymbol)eventSymbol.OriginalDefinition.Type).DelegateInvokeMethod;
@@ -81,7 +80,7 @@ namespace ReactiveMarbles.ObservableEvents.SourceGenerator.EventGenerators.Gener
                 // If we have only one member, produces arguments: (arg1);
                 // If we have greater than one member, produces arguments with value type: ((arg1, arg2))
                 methodParametersArgumentList = invokeMethod.Parameters.Length == 1 ? invokeMethod.Parameters[0].GenerateArgumentList() : invokeMethod.Parameters.GenerateTupleArgumentList();
-                eventArgsType = invokeMethod.Parameters.Length == 1 ? IdentifierName(invokeMethod.Parameters[0].Type.GenerateFullGenericName()) : invokeMethod.Parameters.Select(x => (x.Type, x.Name)).GenerateTupleType();
+                eventArgsType = invokeMethod.Parameters.Length == 1 ? IdentifierName(invokeMethod.Parameters[0].Type.GenerateFullGenericName()) : invokeMethod.Parameters.Select(x => (x.Type, x.Name)).ToList().GenerateTupleType();
             }
             else
             {
@@ -92,61 +91,64 @@ namespace ReactiveMarbles.ObservableEvents.SourceGenerator.EventGenerators.Gener
 
             var eventName = eventSymbol.Name;
 
-            // Produces local function: void Handler(DataType1 eventParam1, DataType2 eventParam2) => eventHandler(eventParam1, eventParam2)
-            var localFunctionExpression = LocalFunctionStatement(
-                                                PredefinedType(Token(SyntaxKind.VoidKeyword)),
-                                                Identifier("Handler"))
-                                            .WithParameterList(invokeMethod.GenerateMethodParameters())
-                                            .WithExpressionBody(
-                                                ArrowExpressionClause(
-                                                    InvocationExpression(IdentifierName("eventHandler"))
-                                                        .WithArgumentList(methodParametersArgumentList)))
-                                            .WithSemicolonToken(Token(SyntaxKind.SemicolonToken));
+            var localFunctionExpression = GenerateLocalHandler(methodParametersArgumentList, invokeMethod, "obs.OnNext");
 
-            // Produces lambda expression: eventHandler => (local function above); return Handler;
-            var conversionLambdaExpression = SimpleLambdaExpression(
-                Parameter(Identifier("eventHandler")),
-                Block(localFunctionExpression, ReturnStatement(IdentifierName("Handler"))));
+            ////// Produces lambda expression: eventHandler => (local function above); return Handler;
+            ////var conversionLambdaExpression = SimpleLambdaExpression(Parameter("eventHandler"), Block(new StatementSyntax[] { localFunctionExpression, ReturnStatement("Handler") }, 3));
 
             // Produces type parameters: <EventArg1Type, EventArg2Type>
-            var fromEventTypeParameters = TypeArgumentList(SeparatedList<TypeSyntax>(new SyntaxNodeOrToken[] { returnType, Token(SyntaxKind.CommaToken), eventArgsType }));
+            var fromEventTypeParameters = new[] { returnType, eventArgsType };
 
-            // Produces: => global::System.Reactive.Linq.Observable.FromEvent<TypeParameters>(h => (handler from above), x => x += DataObject.Event, x => x -= DataObject.Event);
-            var expression = ArrowExpressionClause(
-                InvocationExpression(
-                    MemberAccessExpression(
-                        SyntaxKind.SimpleMemberAccessExpression,
-                        IdentifierName("global::System.Reactive.Linq.Observable"),
-                        GenericName(Identifier("FromEvent"))
-                            .WithTypeArgumentList(fromEventTypeParameters)))
-                        .WithArgumentList(
-                            ArgumentList(
-                                SeparatedList<ArgumentSyntax>(
-                                    new SyntaxNodeOrToken[]
-                                    {
-                                            Argument(conversionLambdaExpression),
-                                            Token(SyntaxKind.CommaToken),
-                                            GenerateArgumentEventAccessor(SyntaxKind.AddAssignmentExpression, eventName, dataObjectName),
-                                            Token(SyntaxKind.CommaToken),
-                                            GenerateArgumentEventAccessor(SyntaxKind.SubtractAssignmentExpression, eventName, dataObjectName)
-                                    }))));
+            var conversionLambdaExpression = SimpleLambdaExpression(Parameter("obs"), Block(
+                new StatementSyntax[]
+                {
+                    localFunctionExpression,
+                    ExpressionStatement(GenerateEventAssignment(SyntaxKind.AddAssignmentExpression, eventName, dataObjectName, "Handler")),
+                    ReturnStatement(InvocationExpression(
+                        MemberAccessExpression(SyntaxKind.SimpleMemberAccessExpression, "global::System.Reactive.Disposables.Disposable", "Create"),
+                        new ArgumentSyntax[]
+                        {
+                            Argument(ParenthesizedLambdaExpression(GenerateEventAssignment(SyntaxKind.SubtractAssignmentExpression, eventName, dataObjectName, "Handler")))
+                        })),
+                },
+                3));
+
+            // Produces: => global::System.Reactive.Linq.Observable.Create<TypeParameter>(obs =>
+            // {
+            //    void Handler(...event params...) => obs.OnNext(params);
+            //    _data.Event += Handler;
+            //    return global::System.Reactive.Disposables.Disposable.Create(() => _data.Event -= Handler);
+            // }
+            var expression = ArrowExpressionClause(InvocationExpression(
+                MemberAccessExpression(SyntaxKind.SimpleMemberAccessExpression, "global::System.Reactive.Linq.Observable", GenericName("Create", new[] { eventArgsType })),
+                new[]
+                {
+                    Argument(conversionLambdaExpression),
+                }));
 
             return (expression, eventArgsType.GenerateObservableType());
         }
 
-        private static ArgumentSyntax GenerateArgumentEventAccessor(SyntaxKind accessor, string eventName, string dataObjectName)
+        private static LocalFunctionStatementSyntax GenerateLocalHandler(IReadOnlyCollection<ArgumentSyntax> methodParametersArgumentList, IMethodSymbol invokeMethod, string handlerName) =>
+
+            // Produces local function: void Handler(DataType1 eventParam1, DataType2 eventParam2) => eventHandler(eventParam1, eventParam2)
+            LocalFunctionStatement(
+                "void",
+                "Handler",
+                invokeMethod.GenerateMethodParameters(),
+                ArrowExpressionClause(
+                        InvocationExpression(handlerName, methodParametersArgumentList)));
+
+        private static AssignmentExpressionSyntax GenerateEventAssignment(SyntaxKind accessor, string eventName, string dataObjectName, string handlerName)
         {
             // This produces "x => dataObject.EventName += x" and also "x => dataObject.EventName -= x" depending on the accessor passed in.
-            return Argument(
-                SimpleLambdaExpression(
-                    Parameter(Identifier("x")),
-                    AssignmentExpression(
+            return AssignmentExpression(
                         accessor,
                         MemberAccessExpression(
                             SyntaxKind.SimpleMemberAccessExpression,
-                            IdentifierName(dataObjectName),
-                            IdentifierName(eventName)),
-                        IdentifierName("x"))));
+                            dataObjectName,
+                            eventName),
+                        handlerName);
         }
     }
 }
