@@ -1,5 +1,5 @@
-﻿// Copyright (c) 2019-2021 ReactiveUI Association Inc. All rights reserved.
-// ReactiveUI Association Inc licenses this file to you under the MIT license.
+﻿// Copyright (c) 2019-2021 ReactiveUI Association Incorporated. All rights reserved.
+// ReactiveUI Association Incorporated licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for full license information.
 
 using System;
@@ -10,20 +10,20 @@ using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 
-using static Microsoft.CodeAnalysis.CSharp.SyntaxFactory;
+using static ReactiveMarbles.ObservableEvents.SourceGenerator.SyntaxFactoryHelpers;
 
 namespace ReactiveMarbles.ObservableEvents.SourceGenerator.EventGenerators.Generators
 {
     internal class StaticEventGenerator : EventGeneratorBase
     {
         /// <inheritdoc />
-        public override NamespaceDeclarationSyntax? Generate(INamedTypeSymbol item, bool generateEmpty)
+        public override NamespaceDeclarationSyntax? Generate(INamedTypeSymbol item)
         {
             var eventWrapperMembers = new List<PropertyDeclarationSyntax>();
 
             var namespaceName = item.ContainingNamespace.ToDisplayString(RoslynHelpers.SymbolDisplayFormat);
 
-            foreach (var eventDetail in item.GetEvents())
+            foreach (var eventDetail in item.GetEvents(true))
             {
                 var eventWrapper = GenerateEventWrapperObservable(eventDetail, item.GenerateFullGenericName(), item.Name);
 
@@ -35,13 +35,16 @@ namespace ReactiveMarbles.ObservableEvents.SourceGenerator.EventGenerators.Gener
 
             if (eventWrapperMembers.Count > 0)
             {
-                var members = ClassDeclaration("RxEvents")
-                    .WithModifiers(TokenList(Token(SyntaxKind.InternalKeyword), Token(SyntaxKind.StaticKeyword)))
-                    .WithLeadingTrivia(XmlSyntaxFactory.GenerateSummarySeeAlsoComment("A class that contains extension methods to wrap events contained within static classes within the {0} namespace.", namespaceName))
-                    .WithMembers(List<MemberDeclarationSyntax>(eventWrapperMembers.Where(x => x != null).Select(x => x!)));
+                var members = new[]
+                {
+                    ClassDeclaration(
+                        "RxEvents",
+                        new[] { SyntaxKind.InternalKeyword, SyntaxKind.StaticKeyword },
+                        eventWrapperMembers.Where(x => x != null).Select(x => (MemberDeclarationSyntax)x!).ToList(),
+                        1)
+                };
 
-                return NamespaceDeclaration(IdentifierName(namespaceName))
-                    .WithMembers(SingletonList<MemberDeclarationSyntax>(members));
+                return NamespaceDeclaration(namespaceName, members, true);
             }
 
             return null;
