@@ -1,5 +1,5 @@
-﻿// Copyright (c) 2019-2021 ReactiveUI Association Inc. All rights reserved.
-// ReactiveUI Association Inc licenses this file to you under the MIT license.
+﻿// Copyright (c) 2019-2021 ReactiveUI Association Incorporated. All rights reserved.
+// ReactiveUI Association Incorporated licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for full license information.
 
 using System;
@@ -10,7 +10,7 @@ using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 
-using static Microsoft.CodeAnalysis.CSharp.SyntaxFactory;
+using static ReactiveMarbles.ObservableEvents.SourceGenerator.SyntaxFactoryHelpers;
 
 namespace ReactiveMarbles.ObservableEvents.SourceGenerator.EventGenerators
 {
@@ -23,35 +23,25 @@ namespace ReactiveMarbles.ObservableEvents.SourceGenerator.EventGenerators
 
         private static MethodDeclarationSyntax GenerateStaticMethod(INamedTypeSymbol declarationType)
         {
-            var eventsClassName = IdentifierName("global::" + declarationType.ContainingNamespace.ToDisplayString(RoslynHelpers.SymbolDisplayFormat) + ".Rx" + declarationType.Name + "Events");
-            return MethodDeclaration(eventsClassName, Identifier("Events"))
-                .WithModifiers(TokenList(Token(SyntaxKind.PublicKeyword), Token(SyntaxKind.StaticKeyword)))
-                .WithParameterList(ParameterList(SingletonSeparatedList(
-                    Parameter(Identifier("item"))
-                        .WithModifiers(TokenList(Token(SyntaxKind.ThisKeyword)))
-                        .WithType(IdentifierName(declarationType.GenerateFullGenericName())))))
-                .WithExpressionBody(ArrowExpressionClause(
-                    ObjectCreationExpression(eventsClassName)
-                        .WithArgumentList(ArgumentList(SingletonSeparatedList(Argument(IdentifierName("item")))))))
-                .WithSemicolonToken(Token(SyntaxKind.SemicolonToken))
-                .WithObsoleteAttribute(declarationType)
+            var eventsClassName = "global::" + declarationType.ContainingNamespace.ToDisplayString(RoslynHelpers.SymbolDisplayFormat) + ".Rx" + declarationType.Name + "Events";
+            var modifiers = new[] { SyntaxKind.PublicKeyword, SyntaxKind.StaticKeyword };
+            var parameters = new[] { Parameter(declarationType.GenerateFullGenericName(), "item", new[] { SyntaxKind.ThisKeyword }) };
+            var body = ArrowExpressionClause(ObjectCreationExpression(eventsClassName, new[] { Argument("item") }));
+            var attributes = RoslynHelpers.GenerateObsoleteAttributeList(declarationType);
+
+            return MethodDeclaration(attributes, modifiers, eventsClassName, "Events", parameters, 0, body)
                 .WithLeadingTrivia(XmlSyntaxFactory.GenerateSummarySeeAlsoComment("A wrapper class which wraps all the events contained within the {0} class.", declarationType.GetArityDisplayName()));
         }
 
         private static MethodDeclarationSyntax GenerateInstanceMethod(INamedTypeSymbol declarationType)
         {
-            var eventsClassName = IdentifierName("global::" + declarationType.ContainingNamespace.ToDisplayString(RoslynHelpers.SymbolDisplayFormat) + ".Rx" + declarationType.Name + "Events");
-            return MethodDeclaration(eventsClassName, Identifier("Events"))
-                .WithModifiers(TokenList(Token(SyntaxKind.PublicKeyword), Token(SyntaxKind.StaticKeyword)))
-                .WithParameterList(ParameterList(SingletonSeparatedList(
-                    Parameter(Identifier("item"))
-                        .WithModifiers(TokenList(Token(SyntaxKind.ThisKeyword)))
-                        .WithType(IdentifierName(declarationType.GenerateFullGenericName())))))
-                .WithExpressionBody(ArrowExpressionClause(
-                    ObjectCreationExpression(eventsClassName)
-                        .WithArgumentList(ArgumentList(SingletonSeparatedList(Argument(IdentifierName("item")))))))
-                .WithSemicolonToken(Token(SyntaxKind.SemicolonToken))
-                .WithObsoleteAttribute(declarationType)
+            var eventsClassName = "global::" + declarationType.ContainingNamespace.ToDisplayString(RoslynHelpers.SymbolDisplayFormat) + ".Rx" + declarationType.Name + "Events";
+            var modifiers = new[] { SyntaxKind.PublicKeyword, SyntaxKind.StaticKeyword };
+            var parameters = new[] { Parameter(declarationType.GenerateFullGenericName(), "item", new[] { SyntaxKind.ThisKeyword }) };
+            var body = ArrowExpressionClause(ObjectCreationExpression(eventsClassName, new[] { Argument("item") }));
+            var attributes = RoslynHelpers.GenerateObsoleteAttributeList(declarationType);
+
+            return MethodDeclaration(attributes, modifiers, eventsClassName, "Events", parameters, 0, body)
                 .WithLeadingTrivia(XmlSyntaxFactory.GenerateSummarySeeAlsoComment("A wrapper class which wraps all the events contained within the {0} class.", declarationType.GetArityDisplayName()));
         }
     }
